CINEPLEXFINAL/backend/requirements.txt
fastapi==0.115.11
httpx==0.28.1
uvicorn==0.27.1
jose==1.0.0
numpy==2.2.4
pandas==2.2.3
pydantic==2.10.6
pydantic_settings==2.8.1
python-dotenv==1.0.1
python_bcrypt==0.3.2
python_jose==3.4.0
Requests==2.32.3
scikit_learn==1.6.1
SQLAlchemy==2.0.39
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy<2.0.0,>=1.4.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
email-validator==1.1.3
Werkzeug==2.0.1
